refactor(team): extract upgrade endpoints into broker_upgrade.go
Slice 1 of broker.go decomposition. Moves handleVersion,
handleUpgradeCheck, handleUpgradeChangelog, handleUpgradeRun,
runUpgradeCmd, tailBuffer (+ Write/Bytes/Truncated),
detectUpgradeInstall, detectLocalInstall, pkgDeclaresWuphf,
withTruncationSentinel, upgradeChangelogKey, upgradeCheckCached,
upgradeChangelogCached, plus the upgradeRun* and upgradeCheck*
caching state, into a new broker_upgrade.go.

No behavior change — pure file relocation within package team.
broker.go shrinks from 12058 to 11424 lines. Adds
TestHandleVersion_ReturnsBuildInfoJSON characterization test:
handleVersion went from 0% to 100% line coverage. Total package
coverage holds at 62.9% baseline.
